moved cerberus reasons to a dedicated file
